fix: Bug fixes in delete users feature

+++ headers/decor.h
@@ -16,7 +16,7 @@ void textGreen();
 void textRed();
 void textBlue();
 void textYellow();
-void alert(char *message, int time);
+void alert(char *message, float time);
 void command(char *command);
 void stripNewLine(char *str);
